+++ semantic_goals_generator/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package semantic_goals_generator
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.0.0 (18-02-2022)
+------------------
+* Change orientation subtopic to direction.
+* Added yaw in srv message and in configuration file. 
+
 1.0.1 (15-12-2021)
 ------------------
 * Update simple_laser_geometry to 3.0.0.